functions, first pass
Left lots of TODOs in here. Lots of messy code that will get cleaned up
... shortly.

Change lox numbers from f32 to f64 to get enough precision for time.

+++ src/token.rs
@@ -39,7 +39,7 @@ pub enum TokenData {
 
     Identifier(String),
     StringToken(String),
-    Number(f32),
+    Number(f64),
 
     // Keywords.
     And,
